Extract findVisibleCandidates() et. al.  from functionResolution.cpp
